k8s-app/backend/requirements.txt
# Python dependencies
Flask==2.3.2
mysql-connector-python==8.0.33
